==> 1_pre-EDA.ipynb <==
# Análisis y predicciones de la seguridad vial en Reino Unido
***Cristian Escudero***

El proyecto pone de manifiesto una serie de datos que hacen referencia a accidentes de tráfico ocurridos en Reino Unido durante el período de 1979 - 2004.

Los datos han sido proporcionados en por el gobierno de Reino Unido. [(Datos)](https://academictorrents.com/details/c7d2d7a91ae3fd0256dd2ba2d7344960cb3c4dbb)

Se analizarán 3 datasets de datos, sobre accidentes, víctimas y vehículos implicados en el cual se descatarán las características principales encontradas.

A continuación se realizará un modelo predictivo para poder predecir la gravedad "Fatal" de accidente y poder así encontrar los puntos críticos donde se acumulan.

Para ello el proyecto será divido en las 5 tareas siguientes:

- Limpieza y transformación de los datos
- Visualización
- Feature Engineering
- Machine learning
- Front End



import os
chdir = os.getcwd()

#### *Guardamos el path de los datos y del repositorio en un txt para usarlo sucesivamente en notebooks posteriores.*

datapath = input("Introduce datasets path")
with open("path.txt",mode="w") as f:
    f.write(datapath)
    f.write("\n")
    f.write(chdir)
datapath = datapath +'/'

datapath

**Importamos librerías**

import plotly.tools as tls
import plotly as py
import plotly.graph_objs as go

from matplotlib import *
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from matplotlib import cm
cmap = cm.get_cmap
import seaborn as sns
import modules

%matplotlib inline
import io
pd.set_option('display.max_columns', None)
from plotly.subplots import make_subplots
import plotly.graph_objects as go
import streamlit as st
import cufflinks as cf
cf.go_offline()
cf.set_config_file(theme='pearl',sharing='public',offline=True)

**Comprobando la codificación del archivo**

import chardet
data1 = datapath+"Accidents7904.csv"
data2 = datapath+"Vehicles7904.csv"
data3 = datapath+"Casualty7904.csv"
files = [data1,data2,data3]
for i in files:
    with open(i, 'rb') as rawdata:
        result = chardet.detect(rawdata.read(10000))
    print(result)

### ***Accidents dataset (1974-2004)***

#### 1. Importación y exploración de los datos

accidents=pd.read_csv(data1,delimiter=',',encoding='UTF-8-SIG',low_memory=False)

accidents.shape

accidents.head()

accidents.info()

accidents.describe().T

En general podemos ver que muchas de las variables en el dataset Accidents tienen una subcategoría en su variable que acumula más porcentaje,por lo tanto podemos asegurar que es un dataset desbalanceado, como por ejemplo las columnas Accident_Severity, Road_type, Junction_Control, Light_Conditions ,Weather_Conditions..., entre otras.

####  2. Limpieza de los datos

**Comprobamos porcentaje de NaN o elementos nulos**

print("NaN data: \n \n"," \n ",accidents.isna()\
                                          .sum()/len(accidents),"%")
print("\n Missing or out of range data:\n",np.abs(accidents[accidents==-1]\
                                            .sum())/len(accidents),"%")
#accidents.isna().sum().sum()/len(accidents)

**Decartamos las columnas que no nos interesan, con alto porcentaje de NaN o valores -1 (missing values)**

columns_to_drop = ['Location_Easting_OSGR', 
                   'Location_Northing_OSGR',
                    'Police_Force',
                   'Local_Authority_(District)',
                   'Local_Authority_(Highway)',
                   '1st_Road_Number',
                   '2nd_Road_Number',
                   'Pedestrian_Crossing-Human_Control', 
                    'Pedestrian_Crossing-Physical_Facilities',
                    'Did_Police_Officer_Attend_Scene_of_Accident',
                    'LSOA_of_Accident_Location','Longitude','Latitude',
                    'Urban_or_Rural_Area',
                   'Junction_Control','2nd_Road_Class',
                    'Special_Conditions_at_Site',
                    'Carriageway_Hazards',
                    'Junction_Detail']

accidents.drop(labels=columns_to_drop,inplace=True,axis=1)

**Descartamos filas con NaN**

accidents.dropna(inplace=True)

The values = -1 are refered to missing values or incorrect data.

**Eliminamos filas con valores -1**

for i,k in enumerate(accidents):
    accidents.drop(index=accidents[accidents[k] == -1]\
            .index, inplace=True)

accidents.shape

#### 3. Procesamiento de los datos

Los datos han sido procesados con el fin de obtener un mejor desempeño en el análisis y evitar los datos sesgados.
 Eso implica reemplazar la categoría numérica por etiquetas originales, generar nuevas columnas y eliminar las columnas que no necesitamos y crear funciones que nos ayuden en este proceso.

*Reemplazo de valores con las etiquetas correctas para las columnas siguientes*:
- Date, Month, Year, Hour,Time
- Daytime
- Light_Conditions
- Road_Surface_Conditions
- Weather_Conditions
- Speed
- Accident_Severity
- Day_of_Week
- 1st_road_class
- Road_Type


**Date, Month, Year, Hour**

Creamos las columnas Month, Year y Hour.

accidents['Date']=pd.to_datetime(accidents['Date'])
accidents['Month']=accidents['Date'].dt.month
accidents['Year']=accidents['Date'].dt.year
accidents['Hour'] = pd.to_datetime(accidents['Time'], format='%H:%M').dt.hour

#Replace int by Months
accidents['Month'] = accidents['Month'].replace([1,2,3,4,5,6,7,8,9,10,11,12],['January', 'February', 
                                                 'March','April', 'May','June',
                                                 'July', 'August', 'September',
                                                'October','November','December'
])
#conver to integer values
accidents['Year'] = accidents['Year'].astype(int)

Descartamos la columna Time

accidents.drop(['Time'],axis=1,inplace=True)

##### *Creamos la columna Daytime con los siguientes valores*
- Morning
- Afternoon
- Rush_Hour
- Evening
- Night



Definimos la función para devolver las horas en grupos o momentos del día.


def daytime(hour):
    if hour >= 5 and hour < 10:
        return "Commuting to work"
    elif hour >= 10 and hour < 15:
        return "Office hours"
    elif hour >= 15 and hour < 19:
        return "Commuting to home"
    elif hour >= 19 and hour < 23:
        return "Evening"
    else:
        return "Night"
    
accidents['Daytime'] = accidents['Hour'].apply(daytime)

accidents['Daytime'].value_counts(normalize=True)

print("\n",accidents['Light_Conditions'].value_counts(normalize=True))
print("\n",accidents['Road_Surface_Conditions'].value_counts(normalize=True))
print("\n",accidents['Weather_Conditions'].value_counts(normalize=True))

Como podemos ver el problema aquí es que muchas de las subcategorías apenas tienen datos e información que nos puedan aportar algo.
Por ello se crean nuevas columnas intentando cambiar un poco la estrategia y acumulando más porcentaje para cada categoría nueva creada.

**Conditions**

#Replacing integer values by labels
accidents.Light_Conditions = accidents.Light_Conditions.replace([1,4,5,6,7], 
                                                      ['Daylight', 
                                                       'Darkness - lights lit', 
                                                       'Darkness - lights unlit', 
                                                       'Darkness - no lighting', 
                                                       'Darkness - lighting unknown'])


accidents.Weather_Conditions = accidents.Weather_Conditions.replace([1,2,3,4,5,6,7,8,9], 
                                                                ['Fine no high winds', 
                                                                 'Raining no high winds', 
                                                                 'Snowing no high winds', 
                                                                 'Fine + high winds', 
                                                                 'Raining + high winds', 
                                                                 'Snowing + high winds', 
                                                                 'Fog or mist', 'Other', 'Unknown', 
                                                                 ])

accidents.Road_Surface_Conditions = accidents.Road_Surface_Conditions.replace([1,2,3,4,5,6,7], 
                                                                                    ['Dry', 
                                                                                     'Wet or damp', 
                                                                                     'Snow', 
                                                                                     'Frost or ice', 
                                                                                     'Flood over 3cm. deep',
                                                                                     'Oil or diesel',
                                                                                     'Mud'])
#define functions to groupby smaller groups
def road_conditions(cond):
    if cond != "Dry":
        return 'Bad'
    elif cond == 'Dry':
        return 'Good'
accidents['Road_Surface_Conditions_2'] = accidents['Road_Surface_Conditions'].apply(road_conditions)

def weather_conditions(value1):
    if value1 == 'Fine no high winds':
        return 'Good'
    else:
        return 'Bad'
accidents['Weather_Conditions_2'] =  accidents.Weather_Conditions.apply(weather_conditions)


def light_conditions(value2):
    if value2 == 'Daylight':
        return 'Good'
    else:
        return 'Bad'
accidents['Light_Conditions_2'] = accidents['Light_Conditions'].apply(light_conditions)

print(accidents['Light_Conditions_2'].value_counts(normalize=True))
print(accidents['Road_Surface_Conditions_2'].value_counts(normalize=True))
print(accidents['Weather_Conditions_2'].value_counts(normalize=True))

Hemos mejorado un poco los porcentajes y reducido las subcategorías, pero aún hay  que muestran un porcentaje mucho más alto, esto podría suponer un problema a la hora de entrenar nuestro modelo.

Eliminamos las columnas originales

accidents.drop(['Light_Conditions',
                'Weather_Conditions',
                'Road_Surface_Conditions'],
               inplace=True,axis=1)

**Speed**

Para la variable Speed hacemos lo mismo, la agrupamos en grupos de velocidades.

def speed_groups(speed):
    if speed < 10:
        return "0-10"
    elif speed >=10 and speed < 20:
        return "10-20"
    elif speed >=20 and speed < 30:
        return "20-30"
    elif speed >=30 and speed < 40:
        return "30-40"
    elif speed >=40 and speed < 50:
        return "40-50"
    elif speed >=50 and speed < 60:
        return "50-60"
    elif speed >=60 and speed < 70:
        return "60-70"
    elif speed >=70 and speed < 80:
        return "70-80"
    elif speed >=80 and speed < 90:
        return "80-90"
    
accidents['Speed_limit_2'] =  accidents['Speed_limit'].apply(speed_groups)
accidents['Speed_limit_2'].value_counts()

**Accident_Severity**

accidents['Accident_Severity'].replace([1,2,3],
                                       ['Fatal',
                                        'Serious',
                                        'Slight']
                                       ,inplace=True)


accidents['Accident_Severity'].value_counts(normalize=True)

**Day_of_Week**

days = ['Sunday',
        'Monday',
        'Tuesday',
        'Wednesday',
        'Thursday',
        'Friday',
        'Saturday']

accidents['Day_of_Week'].replace([1,2,3,4,5,6,7],days, inplace=True)

accidents.Day_of_Week.value_counts(normalize=True)\
                     .sort_values(ascending=True)


**1st_Road_Class**

#Put Motorway and A(M) in the same category
accidents['1st_Road_Class'].replace([1,2,3,4,5,6],
                                    ['Motorway',
                                     'A(M)',
                                     'A',
                                     'B',
                                     'C',
                                     'Unclassified'],inplace=True)

accidents['1st_Road_Class'] = accidents['1st_Road_Class'].replace('A(M)', 'Motorway')
accidents['1st_Road_Class'].value_counts(normalize=True)

**Road_Type**

accidents['Road_Type'].replace([1,2,3,6,7,9,12],
                               ['Roundabout',
                                 'One way street',
                                 'Dual carriageway',
                                 'Single carriageway',
                                 'Slip road','Unknown',
                                 'One way street/Slip road'],
                               inplace=True)
accidents.Road_Type.value_counts(normalize=True)

**Guardamos en csv**

accidents.head()

accidents.to_csv(datapath+'accidents_labeled.csv',sep=',',index = True)

### ***Vehicles and Casualty datasets (1974 - 2004)***

#### 1. Importamos y exploramos los datos

veh = pd.read_csv(data2,
                  delimiter=',',
                  encoding='UTF-8-SIG',
                  usecols=[
                      'Acc_Index','Vehicle_Type',
                      'Sex_of_Driver',
                      'Age_Band_of_Driver',
                      'Engine_Capacity_(CC)',
                      'Age_of_Vehicle'
])

cas = pd.read_csv(data3
                  ,delimiter=',',
                  encoding='UTF-8-SIG',
                  usecols=['Acc_Index',
                           'Sex_of_Casualty',
                           'Age_Band_of_Casualty',
                           'Casualty_Type'])

**Juntamos ambos datasets**

dfmerged = pd.merge(cas,veh,how='inner',on='Acc_Index')

del cas,veh
dfmerged.head()

dfmerged.info()

dfmerged.describe().T

dfmerged.shape

#### 2. Limpieza de datos

**Comprobamos valores nulos y faltantes**

print("NaN data: \n \n"," \n ",dfmerged.isna()\
                                       .sum()/len(dfmerged),"%")

print("\n Missing or out of range data:\n",np.abs(dfmerged[dfmerged==-1]\
                                             .sum())/len(dfmerged),"%")

**Descartamos las filas con valores nulos**

for i,k in enumerate(dfmerged):
    dfmerged.drop(index=dfmerged[dfmerged[k] == -1].index, inplace=True,axis=1)
    
#Drop unknown values of Sex_of_Driver
dfmerged.drop(index=dfmerged[dfmerged['Sex_of_Driver'] == 3].index, inplace=True,axis=1)

#### 3. Procesamiento de datos

*Reemplazo de valores con las etiquetas correctas las las siguientes columnas:*
- Sex_of_Casualty
- Age_Band_of_Casualty
- Casualty_Type
- Vehicle_Type
- Sex_of_Driver
- Age_Band_of_Driver
- Engine Capacity

**Sex_of_Casualty**

dfmerged.Sex_of_Casualty = dfmerged.Sex_of_Casualty\
                                    .replace([1,2],['Male','Female'])

**Age Band of Casualty**

dfmerged.Age_Band_of_Casualty = dfmerged.Age_Band_of_Casualty\
                                        .replace([1,2,3,4,5,6,7,8,9,10,11],['0 - 5',
                                                                            '6 - 10',
                                                                            '11 - 15',
                                                                            '16 - 20',
                                                                            '21 - 25',
                                                                            '26 - 35',
                                                                            '36 - 45',
                                                                            '46 - 55',
                                                                            '56 - 65',
                                                                            '66 - 75',
                                                                            'Over 75' ])

**Casualty type**

values1=['Pedestrian',
'Cyclist',
'Motorcycle 50cc and under rider or passenger',
'Motorcycle 125cc and under rider or passenger',
'Motorcycle over 125cc and up to 500cc rider or  passenger',
'Motorcycle over 500cc rider or passenger',
'Taxi/Private hire car occupant',
'Car occupant',
'Minibus (8 - 16 passenger seats) occupant',
'Bus or coach occupant (17 or more pass seats)',
'Horse rider',
'Agricultural vehicle occupant',
'Tram occupant',
'Van / Goods vehicle (3.5 tonnes mgw or under) occupant',
'Goods vehicle (over 3.5t. and under 7.5t.) occupant',
'Goods vehicle (7.5 tonnes mgw and over) occupant',
'Mobility scooter rider',
'Electric motorcycle rider or passenger',
'Other vehicle occupant',
'Motorcycle - unknown cc rider or passenger',
'Goods vehicle (unknown weight) occupant',
'Motorcycle - Scooter rider or passenger',
'Motorcycle rider or passenger',
'Motorcycle - Combination rider or passenger',
'Motorcycle over 125cc rider or passenger',
'Taxi (excluding private hire cars) occupant',
'Car occupant (including private hire cars)',
'Minibus/Motor caravan occupant',
'Goods vehicle (over 3.5 tonnes) occupant']

values2 = [0,1,2,3,4,5,8,9,
           10,11,16,17,18,19,
           20,21,22,23,90,97
           ,98,103,104,105,106
           ,108,109,110,113]

dfmerged.Casualty_Type = dfmerged.Casualty_Type.replace(values2,values1)

**Vehicle type**

values3 = [1,2,3,10,11,17,19,
           20,21,90,103,104,
           105,106,108,109,
           110,113]

values4 = ['Pedal cycle',
'Motorcycle 50cc and under',
'Motorcycle 125cc and under',
'Minibus (8 - 16 passenger seats)',
'Bus or coach (17 or more pass seats)',
'Agricultural vehicle',
'Van / Goods 3.5 tonnes mgw or under',
'Goods over 3.5t. and under 7.5t',
'Goods 7.5 tonnes mgw and over',
'Other vehicle',
'Motorcycle - Scooter',
'Motorcycle',
'Motorcycle - Combination',
'Motorcycle over 125cc',
'Taxi (excluding private hire cars)',
'Car (including private hire cars)',
'Minibus/Motor caravan',
'Goods vehicle over 3.5 tonnes']






dfmerged.Vehicle_Type = dfmerged.Vehicle_Type.replace(values3,values4)

**Engine Capacity**

En esta columna hemos detectado que tenemos valores erróneos para la subcategoría Cars y Motorcycle, donde hemos encontrado cilindradas de más de 9000cm3 para coches y 2300cm3 para motocicletas.

Se eliminan ya que nos generarán outliers más adelante.

#Cleaning to a common sense engine capacity / vehicle type
mask1 = dfmerged[(dfmerged["Vehicle_Type"].str.contains("Car"))\
                 & (dfmerged["Engine_Capacity_(CC)"] > 9000)].index
dfmerged.drop(mask1,inplace=True)

mask2 = dfmerged[(dfmerged["Vehicle_Type"].str.contains("Motorcycle"))\
                 & (dfmerged["Engine_Capacity_(CC)"] > 2300)].index
dfmerged.drop(mask2,inplace=True)

mask3 = dfmerged[(dfmerged["Vehicle_Type"].str.contains("Taxi"))\
                 & (dfmerged["Engine_Capacity_(CC)"] > 2400)].index
dfmerged.drop(mask3,inplace=True)

mask4 = dfmerged[dfmerged['Vehicle_Type']=="Other vehicle"].index
dfmerged.drop(mask4,inplace=True)

**Sex_of_Driver**

dfmerged.Sex_of_Driver = dfmerged.Sex_of_Driver.replace([1,2],['Male','Female'])

**Age_Band_of_Driver**

dfmerged.Age_Band_of_Driver = dfmerged.Age_Band_of_Driver\
                                      .replace([1,2,3,4,5,6,7,8,9,10,11],['0 - 5','6 - 10','11 - 15',
                                                                        '16 - 20','21 - 25','26 - 35',
                                                                        '36 - 45','46 - 55','56 - 65',
                                                                        '66 - 75','Over 75'
                                                                        ])

**Comprobamos el dataframe procesado**

dfmerged.head()

**Juntamos los dos últimos datasets y lo guardamos como csv**

#Rename the column before merge.
accidents.rename(columns = {'Accident_Index':'Acc_Index'}, inplace = True)

dfmerged.to_csv(datapath+'veh_cas_labeled.csv',sep=',',index = True)

#Merge the Accidents, casualties and vehicles dataframes cleaned from the last notebook.
pd.merge(dfmerged,accidents,how='inner',on = 'Acc_Index').to_csv(datapath+'alldfjoined.csv',sep=',',index = True)

reset -f
